Refactor color tooltip
I can't test this as I don't know what it does and what to look for
